+++ source/typescript/hybrid/process_recognizer_states.ts
@@ -12,12 +12,12 @@ import {
     createAssertionShiftManual,
     rec_state_prod,
     rec_state
-} from "../util/global_names.js";
-import { AS, ExprSC, SC } from "../util/skribble.js";
+} from "../utilities/global_names.js";
+import { AS, ExprSC, SC } from "../utilities/skribble.js";
 import { RenderBodyOptions } from "../types/render_body_options";
 import { processProductionChain } from "../compiler/states/process_production_chain.js";
 import { renderItem, renderProductionCall } from "./item_render_functions.js";
-import { Item } from "../util/item.js";
+import { Item } from "../utilities/item.js";
 import { Symbol, TokenSymbol } from "../types/symbol.js";
 import { RecognizerState, TRANSITION_TYPE } from "../types/recognizer_state.js";
