+++ source/typescript/render/recognizer_template.ts
@@ -1,9 +1,9 @@
 import { Grammar } from "../types/grammar.js";
 import { SymbolType } from "../types/symbol_type";
-import { getProductionFunctionName } from "../util/render_item.js";
+import { getProductionFunctionName } from "../utilities/render_item.js";
 import { action32bit_array_byte_size_default, error8bit_array_byte_size_default, jump16bit_table_byte_size } from "../runtime/parser_memory.js";
 import { RDProductionFunction } from "../types/rd_production_function";
-import { SC } from "../util/skribble.js";
+import { SC } from "../utilities/skribble.js";
 import {
     TokenSpaceIdentifier,
     TokenNumberIdentifier,
@@ -14,8 +14,8 @@ import {
     rec_consume_assert_call,
     rec_glob_lex_name,
     rec_state
-} from "../util/global_names.js";
-import { convertAssertionFunctionBodyToSkribble } from "../util/code_generating.js";
+} from "../utilities/global_names.js";
+import { convertAssertionFunctionBodyToSkribble } from "../utilities/code_generating.js";
 import { createLexerCode } from "./lexer_template.js";
 import { createStateCode } from "./state_template.js";
 import { Helper } from "../compiler/helper.js";
