+++ source/typescript/types/rd_production_function.ts
@@ -1,4 +1,4 @@
-import { SC } from "../util/skribble.js";
+import { SC } from "../utilities/skribble.js";
 export interface RDProductionFunction {
 
     id: number,
