--- source/typescript/util/__global.d.ts
@@ -1,2 +1,0 @@
-declare global {
-}
