+++ source/typescript/utilities/array_globals.d.ts
@@ -1,6 +1,4 @@
 
-import "./__global.js";
-
 declare global {
     interface Array<T> {
         /**
