+++ source/typescript/utilities/closure.ts
