+++ source/typescript/utilities/default_map.ts
