+++ source/typescript/utilities/first.ts
