+++ source/typescript/utilities/follow.ts
