+++ source/typescript/utilities/global_names.ts
