+++ source/typescript/utilities/grammar.ts
