+++ source/typescript/utilities/item_map.ts
