+++ source/typescript/utilities/render_item.ts
