+++ source/typescript/utilities/skribble.ts
