+++ source/typescript/utilities/symbol.ts
