+++ source/typescript/utilities/transition_tree.ts
